-Фикс ошибки с переполнением памяти при формировании отчётов по количеству объектов;
git-tfs-id: [http://192.168.3.86:8080/tfs/defaultcollection]$/CIPJS/KadOzenka;C63375

+++ Source/KadOzenka.Dal/ManagementDecisionSupport/Dto/StatisticalData/NumberOfObjectsByAdministrativeDistrictsByGroupsDto.cs
@@ -8,6 +8,6 @@
 		public string ThirdParentName { get; set; }
 		public string Group { get; set; }
 		public bool HasGroup { get; set; }
-		public int ObjectsCount { get; set; }
+		public long ObjectsCount { get; set; }
 	}
 }
